requirements.txt
appdirs==1.4.0
appnope==0.1.0
coverage==4.3.4
decorator==4.0.11
Django==1.10.5
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-filer==1.2.6
django-mptt==0.8.7
django-polymorphic==1.0.2
easy-thumbnails==2.3
ipython==5.1.0
ipython-genutils==0.1.0
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
Unidecode==0.4.20
wcwidth==0.1.7
